Add simple support for Kleisli[Id, SparkContext, A].
This commit was proposed by Adelbert Chang to make working with
Spark a bit nicer. The reader pattern is designed for exactly
the situation that SparkContext is used in, so we can easily
support it this way.

Examples:

  val st0: SparkTask[A] =
    SparkTask { sc =>
      // code using SparkContext
      ...
    }

  val st1: SparkTask[Int] =
    SparkTask.pure(333)

SparkTask[A] is an alias for Kleisli[Id, SparkContext, A]
along with a companion object (as shown above).